+++ backend/src/main/java/com/ib/alocacao_carros/dto/PagamentoCadastroDTO.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Getter
 public class PagamentoCadastroDTO {
     private String formaPagamento;
-    private Long idLocacao;
+    private Integer idLocacao;
 }
