+++ backend/src/main/java/com/ib/alocacao_carros/repository/CobrancaAdicionalRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CobrancaAdicionalRepository extends JpaRepository<CobrancaAdicional, Long> {
+public interface CobrancaAdicionalRepository extends JpaRepository<CobrancaAdicional, Integer> {
 }
